Refactor Jsx namespace
