+++ src/Esprima/Ast/Jsx/JsxClosingElement.cs
@@ -1,7 +1,6 @@
-﻿using Esprima.Ast;
-using Esprima.Jsx.Utils;
+﻿using Esprima.Utils.Jsx;
 
-namespace Esprima.Jsx.Ast;
+namespace Esprima.Ast.Jsx;
 
 public sealed class JsxClosingElement : JsxExpression
 {
