+++ src/Esprima/Ast/Jsx/JsxNamespacedName.cs
@@ -1,8 +1,7 @@
 ﻿using System.Diagnostics;
-using Esprima.Ast;
-using Esprima.Jsx.Utils;
+using Esprima.Utils.Jsx;
 
-namespace Esprima.Jsx.Ast;
+namespace Esprima.Ast.Jsx;
 
 [DebuggerDisplay("{Namespace,nq}.{Name,nq}")]
 public sealed class JsxNamespacedName : JsxExpression
