+++ src/Esprima/Utils/Jsx/JsxAstJson.cs
@@ -1,8 +1,7 @@
 ﻿using Esprima.Ast;
-using Esprima.Jsx.Ast;
-using Esprima.Utils;
+using Esprima.Ast.Jsx;
 
-namespace Esprima.Jsx.Utils;
+namespace Esprima.Utils.Jsx;
 
 public sealed class JsxAstToJsonConverter : AstToJsonConverter
 {
