+++ src/Esprima/Utils/Jsx/JsxAstVisitor.cs
@@ -1,8 +1,6 @@
-﻿using System.Runtime.CompilerServices;
-using Esprima.Jsx.Ast;
-using Esprima.Utils;
+﻿using Esprima.Ast.Jsx;
 
-namespace Esprima.Jsx.Utils;
+namespace Esprima.Utils.Jsx;
 
 public class JsxAstVisitor : AstVisitor, IJsxAstVisitor
 {
